voevoda/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.1.1
fake-useragent==1.5.1
idna==3.8
lxml==5.3.0
pydantic==2.9.1
pydantic_core==2.23.3
python-dotenv==1.0.1
redis==5.0.8
requests==2.32.3
soupsieve==2.6
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
